refactor: fix migrations 3

+++ database/migrations/2024_10_16_172248_create_embarques_standards_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('op_embarques_standards', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('embarque_id')->unsigned()->index();
-            $table->mediumInteger('standard_id')->unsigned()->index();
+            $table->smallInteger('standard_id')->unsigned()->index();
             $table->foreign('embarque_id')->references('id')->on('op_embarques');
             $table->foreign('standard_id')->references('id')->on('cat_standards');
             $table->timestamps();
